backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
gunicorn==21.2.0
whitenoise==6.7.0
python-dotenv==1.0.1
dj-database-url==2.2.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
